fixed bug in weighted coreset version of poisson cpp code

+++ examples/hilbert_logistic_poisson_regression/stanCppCode/pystan_model_poisson_weighted_coreset_version_2eb5fdf9460ca0b941e7240dfac57ef030aceee3.cpp
@@ -23,7 +23,7 @@ private:
         std::vector<int> y;
         //
         //WEIGHTS MODIFICATION
-        std::vector<int> w;
+        std::vector<double> w;
         //
         //
         matrix_d x;
